Added inches to player height.
