+++ Public/assets/js/main.js
@@ -1,5 +1,5 @@
 import { playerInformation , NBAPlayerGif, NBAPlayerInfo, playerStatistics, seasonStats} from "./playerInfo.js";
-import { deckLoad } from "./deckLoad.js";
+import { deckLoad, newCard, cardClear} from "./cardFunctions.js";
 
 let savedDeck = [];
 let baseStats = [];
@@ -9,6 +9,7 @@ $("#submit").on("click", function () {
   playerInformation();
   $("#SearchPlayer").val('');
   $("#gif, #playerName, #playerID, #playerHeight, #playerweight, #playerteam, #playercity").html("");
+
 });
 
 $("#clear").on("click", function () {
@@ -44,68 +45,15 @@ $("#save-card").on("click", function (event) {
         // console.log(userDeck[n].player_id)
       }
     }
-
   }
+  deckLoad();
 });
-
-let newCard = () => {
-  var newCard = {
-
-    playerID: $("#playerID").text().trim(),
-    playerName: $("#playerName").text().trim(),
-    playerHeight: $("#playerHeight").text().trim(),
-    playerWeight: $("#playerweight").text().trim(),
-    playerTeam: $("#playerteam").text().trim(),
-    playerCity: $("#playercity").text().trim(),
-    playerGIF: $("#plyr_gif").attr('src').trim()
-  };
-  $.post("/api/new", newCard)
-}
-
-// $("#save-card").on("click", function (event) {
-//   event.preventDefault();
-//   let plyrID = $("#playerID").text().trim();
-//   let userDeck = savedDeck[0];
-//   console.log(plyrID)
-//   console.log(userDeck)
-//   if (plyrID === undefined) {
-//     alert("The Card is empty, no card to save");
-//   } else {
-//     for (const n in userDeck) {
-//       console.log(userDeck[n].player_id)
-//       console.log([n].player_id)
-//       if (userDeck[n].player_id === plyrID) {
-//         console.log(userDeck[n])
-//         alert("You already have this card in your deck!");
-//         $("#SearchPlayer").empty()
-//         cardClear();
-//       } else {
-//         // console.log(userDeck[n].player_id)
-//         var newCard = {
-
-//           playerID: $("#playerID").text().trim(),
-//           playerName: $("#playerName").text().trim(),
-//           playerHeight: $("#playerHeight").text().trim(),
-//           playerWeight: $("#playerweight").text().trim(),
-//           playerTeam: $("#playerteam").text().trim(),
-//           playerCity: $("#playercity").text().trim(),
-//           playerGIF: $("#plyr_gif").attr('src').trim()
-//         };
-//         $.post("/api/new", newCard)
-//       }
-//     }
-//     // deckLoad()
-//   }
-// });
 
 $("#VDeck").on("click", function (event) {
   event.preventDefault();
 deckLoad()
 });
 
-let cardClear = () => {
-  $("#gif, #playerName, #playerID, #playerHeight, #playerweight, #playerteam, #playercity").html("");
-};
 
 $(".decklist").on("click", ".deckcard", function (event) {
   event.preventDefault();
@@ -133,5 +81,5 @@ $(".decklist").on("click", ".deckcard", function (event) {
 
   // Enter code to get player season info here
 });
-playerInformation()
+
 deckLoad()
